feat(c303): chamber optimizations — 3 per chamber × 7 chambers
Layout: chamberMeta() helper for unique canonical + OG per chamber
Layout: canonical URL now chamber-specific (was always root /)

Globe:  chamberMeta with geo-specific description and canonical
Pulse:  revalidate=60 (was 300s default)
Pulse:  agent attestations from ECHO_STATE in ledger timeline
Pulse:  MIC provisional ticker in vault section header

Signals: watch/elevated severity left-border color coding on cards
Signals: last-sweep staleness badge (amber >10min, stale label)

Sentinel: confidence floor breach badge (↓ FLOOR at <0.65)
Sentinel: last_journal_at relative time with >15min amber color
Sentinel: QuorumPanel showing sentinel quorum from vault lane

Ledger: source provenance filter strip (github/echo/agent_commit/backfill)
Ledger: filter clears to 'all' on cycle change, shows counts

Vault:  VaultPageClient.tsx extracted from page.tsx (V-1 hybrid SSR prep)
Vault:  QuarantinePanel for quarantined/audit blocks with expand toggle
Vault:  SubstrateAttestationRow with human-readable error + fix hint

Journal: chamberMeta canonical + OG (existing agent filters/cycle tabs preserved)
Sentinel: chamberMeta canonical + OG
Ledger: chamberMeta canonical + OG

Cycle: C-302 → C-303
