+++ app/terminal/journal/page.tsx
@@ -1,9 +1,11 @@
-import type { Metadata } from 'next';
+import { chamberMeta } from '../layout';
 import JournalPageClient from './JournalPageClient';
 
-export const metadata: Metadata = {
-  title: 'Journal · Mobius Terminal',
-};
+export const metadata = chamberMeta(
+  'Journal',
+  'Agent journal entries — inference trails, cycle attestations, and observation records by agent.',
+  'journal'
+);
 
 export default function JournalPage() {
   return <JournalPageClient />;
