+++ app/terminal/ledger/page.tsx
@@ -1,9 +1,11 @@
-import type { Metadata } from 'next';
+import { chamberMeta } from '../layout';
 import LedgerPageClient from './LedgerPageClient';
 
-export const metadata: Metadata = {
-  title: 'Ledger · Mobius Terminal',
-};
+export const metadata = chamberMeta(
+  'Ledger',
+  'EPICON event ledger — MII-rated civic events with verdict tracking, integrity delta, and source provenance.',
+  'ledger'
+);
 
 export default function LedgerPage() {
   return <LedgerPageClient />;
