+++ app/terminal/sentinel/page.tsx
@@ -1,9 +1,11 @@
-import type { Metadata } from 'next';
+import { chamberMeta } from '../layout';
 import SentinelPageClient from './SentinelPageClient';
 
-export const metadata: Metadata = {
-  title: 'Sentinel · Mobius Terminal',
-};
+export const metadata = chamberMeta(
+  'Sentinel',
+  'Agent liveness monitor — heartbeat, confidence, journal cadence, and quorum attestation status.',
+  'sentinel'
+);
 
 export default function SentinelPage() {
   return <SentinelPageClient />;
