+++ app/terminal/signals/page.tsx
@@ -1,9 +1,11 @@
-import type { Metadata } from 'next';
+import { chamberMeta } from '../layout';
 import SignalsPageClient from './SignalsPageClient';
 
-export const metadata: Metadata = {
-  title: 'Signals · Mobius Terminal',
-};
+export const metadata = chamberMeta(
+  'Signals',
+  'Micro-instrument signal sweep across 8 agents — composite GI signal health and anomaly tracking.',
+  'signals'
+);
 
 export default function SignalsPage() {
   return <SignalsPageClient />;
